+++ here-naksha-lib-ext-manager/src/main/java/com/here/naksha/lib/extmanager/ExtensionCache.java
@@ -18,6 +18,8 @@
  */
 package com.here.naksha.lib.extmanager;
 
+import static com.here.naksha.lib.core.NakshaContext.currentContext;
+
 import com.here.naksha.lib.core.IExtensionInit;
 import com.here.naksha.lib.core.INaksha;
 import com.here.naksha.lib.core.SimpleTask;
@@ -68,7 +70,8 @@ public class ExtensionCache {
     List<Future<KVPair<Extension, File>>> futures = extensionConfig.getExtensions().stream()
         .filter(extension -> !this.isLoaderMappingExist(extension))
         .map(extension -> {
-          SimpleTask<KVPair<Extension, File>> task = new SimpleTask<>();
+          SimpleTask<KVPair<Extension, File>> task =
+              new SimpleTask<>(currentContext().withSuperUser(true));
           return task.start(() -> downloadJar(extension));
         })
         .toList();